+++ lab1/lab1/TriangleFactory.cpp
@@ -22,8 +22,8 @@ CTriangleFactory * CTriangleFactory::GetInstance()
 
 std::unique_ptr<IShape> CTriangleFactory::CreateShape(const std::vector<std::string> & inputParts) const
 {
-	Point p1 = Point(std::stod(inputParts[1]), std::stod(inputParts[2]));
-	Point p2 = Point(std::stod(inputParts[3]), std::stod(inputParts[4]));
-	Point p3 = Point(std::stod(inputParts[5]), std::stod(inputParts[6]));
+	Point p1 = Point(std::stoll(inputParts[1]), std::stoll(inputParts[2]));
+	Point p2 = Point(std::stoll(inputParts[3]), std::stoll(inputParts[4]));
+	Point p3 = Point(std::stoll(inputParts[5]), std::stoll(inputParts[6]));
 	return std::make_unique<CTriangle>(p1, p2, p3);
 }
